6 is testing doubting (#16)
* add tests for sized_string

* inject memory allocator for test

* create contextual error message

* create cmake lib for test_utils

* add cmake lib for types, move tests of sized_string in it

* add tests for katas

* add empty tests for execution

* add cmake lib for result_handler

* refactor definitions and calls of function pointers to realloc and calloc

+++ src/execution/execution.c
@@ -32,7 +32,7 @@ kata_status run_kata(const kata_t kata, sized_string_t result_buffer) {
     }
 
     // Read output to result_buffer
-    fread(result_buffer.str, (int)result_buffer.len, 1, fp);
+    fread(result_buffer.str, result_buffer.len, 1, fp);
     int status = pclose(fp);
 
     // Check execution status
